+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Vue on Rails 1.0.0
 
+- Internationalization for components
 - Various component generators and scaffolds
 - Specific-page Vue
 - `vue_component` helper
@@ -9,4 +10,4 @@
 - Jest test configuration
 - Vue UI ready
 - State management with simple store https://vuejs.org/v2/guide/state-management.html
-- Vue on Rails application template using `rails new app -m http://vueonrails.com/vue`
+- Vue on Rails application template using `rails new app -m http://vueonrails.com/vue -d postgresql`